requirements[1].txt
fastapi
uvicorn
pydantic
python-multipart
pdf2image
pytesseract
Pillow
PyMuPDF
opencv-python
PyPDF2